+++ src/main/java/cl/duocucjuancarlos/ecomarketspa/Service/UserService.java
@@ -15,25 +15,39 @@ public class UserService {
     private UserRepository userRepository;
 
 
-    public ResponseEntity<List<UserResponse>> findAllUsers() {
-        return (ResponseEntity<List<UserResponse>>) userRepository.getAllUsers();
+    public List<UserResponse> findAllUsers() {
+        return userRepository.getAllUsers();
     }
 
 
     public UserResponse getUserById(int userId) {
-        return userRepository.getUser(userId);
+        if (userId >= 0 && userId < userRepository.getAllUsers().size()) {
+            return userRepository.getUser(userId);
+        }
+        return null;
     }
 
+    //falta logica de negocio
     public UserResponse addNewUser(UserRequest userRequest) {
-        return userRepository.addUser(userRequest);
+        if (userRequest != null) {
+            return userRepository.addUser(userRequest);
+        }
+        return null;
     }
 
-
+    //falta logica de negocio
     public UserResponse updateUserById(int userId, UserRequest userRequest) {
-        return userRepository.updateUser(userId, userRequest);
+        List<UserResponse>  List = userRepository.getAllUsers();
+        if (userId >= 0 && userId < List.size() ) {
+            return userRepository.updateUser(userId, userRequest);
+        }
+        return null;
     }
 
     public UserResponse deleteUserById(int userId) {
+        if (userRepository.getUser(userId) == null) {
+            return null;
+        }
         return userRepository.deleteUser(userId);
     }
 }
